+++ src/karma-proxy.ts
@@ -23,7 +23,6 @@
 
 import karma = require('karma');
 import Koa from 'koa';
-import portfinder = require('portfinder');
 const proxy = require('koa-proxy');
 import {createServer, Server} from 'http';
 
@@ -54,13 +53,19 @@ interface ConfigFile {
 export type Options = {
   karmaConfig?: karma.ConfigOptions|ConfigFile,
   karmaExitCallback?: (exitCode: number) => void,
+  upstreamProxyAddress?: string,
+  upstreamProxyHostname?: string,
   upstreamProxyPort?: number,
 };
 
 export type Servers = {
+  upstreamProxyAddress: string,
   upstreamProxyServer: Server,
+  upstreamProxyHostname: string,
   upstreamProxyPort: number,
   karmaServer: karma.Server,
+  karmaAddress: string,
+  karmaHostname: string,
   karmaPort: number,
 };
 
@@ -76,15 +81,21 @@ export type Servers = {
 export const start = async(
     upstreamProxyServerFactory: UpstreamProxyServerFactory,
     options?: Options): Promise<Servers> => new Promise((resolve, reject) => {
-  const karmaConfig: ConfigOptions =
-      options && options.karmaConfig as karma.ConfigOptions || {};
+  const karmaConfig: Partial<ConfigOptions&{listenAddress: string}&ConfigFile> =
+      options && options.karmaConfig || {};
   const karmaConfigFile: ConfigFile = karmaConfig as ConfigFile;
   const startingUpstreamProxyPort: number =
       options && options.upstreamProxyPort || 9876;
+  const maxUpstreamProxyPort = 65535;
+  const upstreamProxyHostname: string =
+      options && options.upstreamProxyHostname || 'localhost';
+  const upstreamProxyAddress: string =
+      options && options.upstreamProxyAddress || '0.0.0.0';
   if (karmaConfigFile.configFile) {
     const {configFile} = karmaConfigFile;
-    const configSetter = karma.config.parseConfig(configFile, karmaConfig);
-    configSetter.set(karmaConfig);
+    const configSetter =
+        karma.config.parseConfig(configFile, karmaConfig as ConfigOptions);
+    configSetter.set(karmaConfig as ConfigOptions);
   }
 
   const karmaExitCallback: ((exitCode: number) => void)|undefined =
@@ -98,10 +109,14 @@ export const start = async(
   // to a chosen upstream port becoming unavailable when we try to bind to it.
   // Whenever the error is related to the port, we'll retry starting the server.
   // Otherwise, we'll reject the promise with the error encountered.
-  const retryOrReject = (port: number, err: Error) => {
+  const retryOrReject = (err: Error) => {
     const {code} = err as Error & {code: string};
     if (code === 'EADDRINUSE' || code === 'EACCES') {
-      return startUpstreamProxyServer(port + 1);
+      if (lastUpstreamProxyPortTried < maxUpstreamProxyPort) {
+        ++lastUpstreamProxyPortTried;
+        return startUpstreamProxyServer();
+      }
+      return reject(err);
     } else {
       return reject(err);
     }
@@ -118,25 +133,23 @@ export const start = async(
                        async (ctx: unknown, next: unknown) =>
                            await karmaProxyMiddleware(ctx, next))
                        .callback());
-  upstreamProxyServer.on(
-      'error', (err: Error) => retryOrReject(lastUpstreamProxyPortTried, err));
+  upstreamProxyServer.on('error', (err: Error) => retryOrReject(err));
   upstreamProxyServer.on('listening', () => {
     startKarmaServer(lastUpstreamProxyPortTried);
   });
 
-  const startUpstreamProxyServer = (startingPort: number) => {
-    portfinder.getPort(
-        {port: startingPort}, (err: Error, upstreamProxyPort: number) => {
-          if (err) {
-            return retryOrReject(startingPort, err);
-          }
-          try {
-            lastUpstreamProxyPortTried = upstreamProxyPort;
-            upstreamProxyServer.listen(upstreamProxyPort);
-          } catch (err) {
-            return retryOrReject(upstreamProxyPort, err);
-          }
-        });
+  const startUpstreamProxyServer = (): void => {
+    if (typeof lastUpstreamProxyPortTried === 'undefined') {
+      lastUpstreamProxyPortTried = startingUpstreamProxyPort;
+    } else {
+      ++lastUpstreamProxyPortTried;
+    }
+    try {
+      upstreamProxyServer.listen(
+          lastUpstreamProxyPortTried, upstreamProxyAddress);
+    } catch (err) {
+      retryOrReject(err);
+    }
   };
 
   const startKarmaServer = (upstreamProxyPort: number) => {
@@ -146,8 +159,9 @@ export const start = async(
 
     // This bit is important when starting the karma server
     // because if it is opening browsers, it needs to open them
-    // on the upstream port instead of the default karma server
-    // port.
+    // on the upstream host and port instead of the karma server
+    // host and port.
+    karmaConfig.upstreamProxy.hostname = upstreamProxyHostname;
     karmaConfig.upstreamProxy.port = upstreamProxyPort;
 
     const karmaServer = new karma.Server(karmaConfig, karmaExitCallback);
@@ -162,15 +176,25 @@ export const start = async(
     // placeholder.
     karmaServer.on('listening', (karmaPort: number) => {
       const karmaHostname = karmaConfig.hostname || 'localhost';
+      const karmaAddress = karmaConfig.listenAddress || '0.0.0.0';
       const karmaProtocol = karmaConfig.protocol || 'http:';
       karmaProxyMiddleware =
           proxy({host: `${karmaProtocol}//${karmaHostname}:${karmaPort}/`});
-      resolve({upstreamProxyPort, upstreamProxyServer, karmaPort, karmaServer});
+      resolve({
+        upstreamProxyAddress,
+        upstreamProxyHostname,
+        upstreamProxyPort,
+        upstreamProxyServer,
+        karmaAddress,
+        karmaHostname,
+        karmaPort,
+        karmaServer
+      });
     });
 
     karmaServer.start();
     karmaServer.on('close', () => upstreamProxyServer.close());
   };
 
-  startUpstreamProxyServer(startingUpstreamProxyPort);
+  startUpstreamProxyServer();
 });