+++ CHANGELOG.md
@@ -10,7 +10,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
    Unreleased section, uncommenting the header as necessary.
 -->
 
-<!-- ## Unreleased -->
+## Unreleased 
+
+ - Added a `--proxyAddress` and `--proxyHostname` CLI flags
+ - Added an `upstreamProxyAddress` and `upstreamProxyHostname` options in `start()`
+ - Include `karmaAddress`, `karmaHostname`, `upstreamProxyAddress` and `upstreamProxyHostname` in `start()` async response.
+
 <!-- Add new unreleased items here -->
 
 ## [1.0.0-pre.4] - 2019-08-08